requirements.txt
aiohttp>=3.9.0
websockets>=12.0
apscheduler>=3.10.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
openai>=1.30.0
MetaTrader5>=5.0.45
